+++ client/src/modules/StudentDashboard/components/MentorCard/index.tsx
@@ -1,3 +1,1 @@
 export * from './MentorCard';
-export * from './MentorInfo';
-export * from './SubmitTaskSolution';
